% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susy3dapp-package.R
\docType{package}
\name{susy3dapp-package}
\alias{susy3dapp}
\alias{susy3dapp-package}
\title{susy3dapp: SusySafe 3d}
\description{
Demo for the SusySafe 3d App.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zeta-r/susy3dapp}
  \item Report bugs at \url{https://github.com/zeta-r/susy3dapp/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{CorradoLanera@zetaresearch.com} [copyright holder]

Other contributors:
\itemize{
  \item Marco Ghidina \email{MarcoGhidina@zetaresearch.com} [contributor]
  \item Zeta Research [copyright holder]
}

}
\keyword{internal}
